+++ include/core/GameWindow.h
@@ -8,13 +8,13 @@ struct SDL_Window;
 class GameWindow
 {
 public:
-	bool			init(const std::string& windowTitle, int width, int height);
+	bool			init(const std::string& windowTitle, uint16_t width, uint16_t height);
 	void			clean();
 
 	SDL_Window*		getWindowPtr() const { return m_pWindow; }
 
-	int				getWindowWidth() const { return m_windowWidth; };
-	int				getWindowHeight() const { return m_windowHeight; };
+	uint16_t				getWindowWidth() const { return m_windowWidth; };
+	uint16_t				getWindowHeight() const { return m_windowHeight; };
 
 	static GameWindow*	Instance() // Get instance
 	{
@@ -28,8 +28,8 @@ private:
 
 	SDL_Window* m_pWindow;
 
-	int m_windowWidth;
-	int m_windowHeight;
+	uint16_t m_windowWidth;
+	uint16_t m_windowHeight;
 
 	GameWindow() :m_pWindow(nullptr), m_windowWidth(0), m_windowHeight(0) {};				// Prevent outside unwanted construction
 	GameWindow(const GameWindow&) :m_pWindow(nullptr), m_windowWidth(0), m_windowHeight(0) {};	// Prevent construction by copying
