+++ src/core/GameWindow.cpp
@@ -8,7 +8,7 @@ GameWindow* GameWindow::s_pInstance = nullptr;
 /// <summary>
 /// Initializes the Window systems
 /// </summary>
-bool GameWindow::init(const std::string& windowTitle, int width, int height)
+bool GameWindow::init(const std::string& windowTitle, uint16_t width, uint16_t height)
 {
 	m_pWindow = SDL_CreateWindow(windowTitle.c_str(), width/2, height/4, width, height, NULL);
 	if (!m_pWindow)
